REF: Refactoring metadata package This is a combination of 4 commits. ORG: Adding metadata package
STY: conforming to checklist standard

ORG: Shuffling around testing modules

FIX: Broken imports
