+++ skbio/metadata/__init__.py
@@ -1,0 +1,17 @@
+"""
+Metadata (:mod:`skbio.metadata`)
+
+.. currentmodule:: skbio.metadata
+
+This module provides classes for storing and working with metadata.
+"""
+
+# ----------------------------------------------------------------------------
+# Copyright (c) 2013--, scikit-bio development team.
+#
+# Distributed under the terms of the Modified BSD License.
+#
+# The full license is in the file COPYING.txt, distributed with this software.
+# ----------------------------------------------------------------------------
+
+from __future__ import absolute_import, division, print_function
