+++ skbio/sequence/tests/test_sequence.py
@@ -28,8 +28,8 @@ from skbio import Sequence, DNA
 from skbio.util import assert_data_frame_almost_equal
 from skbio.sequence._sequence import (_single_index_to_slice, _is_single_index,
                                       _as_slice_if_single_index)
-from skbio.util._testing import (ReallyEqualMixin, MetadataMixinTests,
-                                 PositionalMetadataMixinTests)
+from skbio.metadata._testing import (ReallyEqualMixin, MetadataMixinTests,
+                                     PositionalMetadataMixinTests)
 
 
 class SequenceSubclass(Sequence):
